// src/ipc/generated/IdleSettings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type IdleSettings = { blink: boolean, saccade: boolean, lookAt: boolean, breath: boolean, springBone: boolean, 
/**
 * 感情に応じた姿勢と待機中の動き (要件 F-14-5)。
 *
 * 後から足した項目なので default が要る。無いと既存のプロファイルが
 * 読めなくなる。
 */
motion: boolean, };